KEYCLOAK-17695 Split MapStorage provider and provider factory
