+++ model/map/src/main/resources/META-INF/services/org.keycloak.models.map.storage.MapStorageProviderFactory
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.keycloak.models.map.storage.chm.ConcurrentHashMapStorageProvider
+org.keycloak.models.map.storage.chm.ConcurrentHashMapStorageProviderFactory
