+++ server-spi-private/src/main/java/org/keycloak/sessions/AuthenticationSessionSpi.java
@@ -26,6 +26,8 @@ import org.keycloak.provider.Spi;
  */
 public class AuthenticationSessionSpi implements Spi {
 
+    public static final String PROVIDER_ID = "authenticationSessions";
+
     @Override
     public boolean isInternal() {
         return true;
@@ -33,7 +35,7 @@ public class AuthenticationSessionSpi implements Spi {
 
     @Override
     public String getName() {
-        return "authenticationSessions";
+        return PROVIDER_ID;
     }
 
     @Override
